It seems like there are changes made to three files in the repository: `resources/views/frontend/home.blade.php`, `resources/views/frontend/contact.blade.php`, and `config/projects.php`. Based on these changes, a suitable commit message in the conventional commit convention could be:
```
feat(frontend): Update home and contact pages and add new info in projects config (main)
```

This message indicates that a new feature (`feat`) has been added to the frontend section of the project (`frontend`). The changes include updating the home and contact pages, as well as adding new information to the projects configuration file. The branch name `main` is used to indicate the issue number, but since the issue number is not explicitly mentioned in the provided information, it's assumed that the changes are related to the main branch.